Save button will apply changes to the database

+++ src/components/Home.js
@@ -5,20 +5,26 @@ import UploadImage from './UploadImage';
 import { connect } from 'react-redux';
 import { changeOrder, updatePhotosAction } from '../actions/photos-actions';
 import axios from 'axios';
+import "react-loader-spinner/dist/loader/css/react-spinner-loader.css"
+import Loader from 'react-loader-spinner'
+
+
 
 export class Home extends Component {
     state = {
         selected: 'portrait',
+        isSaving: false
     }
 
     constructor(props) {
         super(props);
             this.changeSelected = this.changeSelected.bind(this);
             this.saveAlbum = this.saveAlbum.bind(this);
-            this.deleteItem = this.deleteItem.bind(this);
         }
     
     changeSelected(select, index, direction) {
+        this.setState({ selected: select });
+
         const API_GET_PHOTOS = new Promise((resolve, reject) => {
             
             //Make the call 
@@ -40,43 +46,31 @@ export class Home extends Component {
         )
     }
 
-    deleteItem(index) {
-        // This will remove the image and update the store with the action onchangeorder
-        // This will not remove the image from mongodb or cloudinary
-        let newArray = [...this.props.photos];
-        newArray.splice(index, 1);        
-        this.props.onchangeOrder(newArray);
-
-    }   
-
     saveAlbum() {
-        let newArray = []
-        this.props.photos.forEach(element => {
-            newArray.push({image_url: element.image_url, album: element.album})
-        });
-        JSON.stringify(newArray)
+        if(this.props.photos.length >= 1) {
+            this.setState({isSaving: true})
+            let newArray = []
+            let index = 0
+            this.props.photos.forEach(element => {
+                newArray.push({image_url: element.image_url, album: element.album, order: index++})
+            });
+            JSON.stringify(newArray)
 
-
-        axios.delete(`http://localhost:3000/delete/` + this.state.selected)
-        .then(function (response) {
-            console.log(response);
-
-            axios.post('http://localhost:3000/savealbum', newArray, {
-                headers: {
-                  'content-type': 'application/json'
-                }
-              })
-            .then(res => {
-                console.log(res.data);
-            })
-            .catch(err => console.log(err))
-
-        })
-            .catch(function (error) {
-            console.log(error);
-        });
+            axios.delete(`http://localhost:3000/delete/` + this.state.selected)
+                .then(response => {
+                    axios.post('http://localhost:3000/savealbum', newArray, {headers: {'content-type': 'application/json'}}) 
+                })
+                .then(response => {
+                    setTimeout(() => {
+                        this.setState({isSaving: false})
+                    }, 600)
+                    
+                })
+                .catch(err => {
+                    console.log(err);
+                });
+        }
     }
-
 
     render() {
 
@@ -91,9 +85,24 @@ export class Home extends Component {
                     <button onClick={() => this.changeSelected('recent')}>Recent</button>
                 </div>
                 <div className="home__content">
-                    <ImageList category={this.state.selected} deleteItem={this.deleteItem}/>
-                    <UploadImage localUpload={this.localUpload} category={this.state.selected}/>
-                    <button onClick={() => this.saveAlbum()} className="home__content__save-button">Save Album</button>
+
+                    {this.state.isSaving ? 
+                        (<div className="home__content__loader">
+                            <Loader
+                            type="TailSpin"
+                            color="black"
+                            height={100}
+                            width={100}/>
+                            
+                        </div>
+                        ) :
+                    (<div>
+                        <ImageList category={this.state.selected} deleteItem={this.deleteItem}/>
+                        <UploadImage localUpload={this.localUpload} category={this.state.selected}/>
+                        <button onClick={() => this.saveAlbum()} className="home__content__save-button">Save Album</button>
+                    </div>)
+                    }
+                    
                 </div>
             </div>
         )
